# openapi.yml
openapi: 3.0.3
info:
  title: Swagger API de gestão de processos seletivos - OpenAPI 3.0
  description: |-
    Está é uma API Rest que tem como objetivo expor endpoints para realizar a gestão de um processo seletivo.
    
    Essa API Rest faz parte de um trabalho de conclusão de curso de Sistemas de Informação do aluno Lucas Moraes Cavalcante na Universidade Veiga de Almeida.
  termsOfService: https://github.com/LMoraes7/recruitment-selection
  contact:
    name: Lucas Moraes Cavalcante
    email: lukaocavalcante93@gmail.com
  version: 1.0.11
servers:
  - url: http://localhost:8080
...
paths:
...
  '/step/theorical_test':
    post:
      tags:
        - step
      summary: Solicitar criação de uma etapa de teste teórico
      description: |-
        Endpoint responsável por criar uma etapa de teste teórico.
        <br /> <br />
        **Scope:** FUNC_CREATE_STEP
      operationId: createStepTheoricalStep
      parameters:
        - $ref: '#/components/parameters/AuthorizationHeader'
      requestBody:
        required: true
        description: Dados para realizar a criação de uma etapa de teste teórico.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStepTheoricalTestRequest'
            examples:
              createStepTheoricalTestRequest:
                $ref: '#/components/examples/CreateStepTheoricalTestRequest'
      responses:
        '201':
          description: Solicitação efetuada com sucesso.
        '401':
          description: Solicitação não autorizada devido a falta de autenticação.
        '403':
          description: Solicitação não autorizada devido a falta de permissão.
        '422':
          description: A requisição não pode ser processada devido a um erro de negócio.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
              examples:
                questionsInvalidErrorResponse:
                  $ref: '#/components/examples/QuestionsInvalidErrorResponse'
                questionsNotSameTypeErrorResponse:
                  $ref: '#/components/examples/QuestionsNotSameTypeErrorResponse'
        '500':
          description: Ocorreu um erro interno no servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
              examples:
                internalErrorResponse:
                  $ref: '#/components/examples/InternalErrorResponse'
  '/step/external':
    post:
      tags:
        - step
      summary: Solicitar criação de uma etapa externa
      description: |-
        Endpoint responsável por criar uma etapa externa.
        <br /> <br />
        **Scope:** FUNC_CREATE_STEP
      operationId: createStepExternal
      parameters:
        - $ref: '#/components/parameters/AuthorizationHeader'
      requestBody:
        required: true
        description: Dados para realizar a criação de uma etapa externa.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStepExternalRequest'
            examples:
              createStepExternalRequest:
                $ref: '#/components/examples/CreateStepExternalRequest'
      responses:
        '201':
          description: Solicitação efetuada com sucesso.
        '401':
          description: Solicitação não autorizada devido a falta de autenticação.
        '403':
          description: Solicitação não autorizada devido a falta de permissão.
        '500':
          description: Ocorreu um erro interno no servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
              examples:
                internalErrorResponse:
                  $ref: '#/components/examples/InternalErrorResponse'
  '/step/upload_files':
    post:
      tags:
        - step
      summary: Solicitar criação de uma etapa de envio de arquivos
      description: |-
        Endpoint responsável por criar uma etapa de envio de arquivos.
        <br /> <br />
        **Scope:** FUNC_CREATE_STEP
      operationId: createStepUploadFiles
      parameters:
        - $ref: '#/components/parameters/AuthorizationHeader'
      requestBody:
        required: true
        description: Dados para realizar a criação de uma etapa de envio de arquivo.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStepUploadFilesRequest'
            examples:
              createStepUploadFilesRequest:
                $ref: '#/components/examples/CreateStepUploadFilesRequest'
      responses:
        '201':
          description: Solicitação efetuada com sucesso.
        '401':
          description: Solicitação não autorizada devido a falta de autenticação.
        '403':
          description: Solicitação não autorizada devido a falta de permissão.
        '500':
          description: Ocorreu um erro interno no servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
              examples:
                internalErrorResponse:
                  $ref: '#/components/examples/InternalErrorResponse'
components:
  schemas:
...
    DefaultErrorResponse:
      type: object
      description: Dados informando algum erro retornado para o cliente.
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Campo informando o código de erro que aconteceu.
          nullable: false
        message:
          type: string
          description: Mensagem para dar maior detalhamento do erro ocorrido.
          nullable: false
        errors:
          type: array
          description: Informações sobre os campos que causaram o erro.
          nullable: true
          items:
            type: object
            required:
              - field
              - message
            properties:
              field:
                type: string
                description: Nome do campo que causou o erro gerado.
                nullable: false
              message:
                type: string
                description: Mensagem para dar maior detalhamento do erro ocorrido no campo.
                nullable: false
...
    CreateStepTheoricalTestRequest:
      type: object
      description: Dados para solicitar a crição de um processo seletivo.
      required:
        - title
        - description
        - questions_identifiers
      properties:
        title:
          type: string
          description: Título da etapa.
          nullable: false
        description:
          type: string
          description: Descrição da etapa.
          nullable: false
        questions_identifiers:
          type: array
          description: Identificadores de questões a serem utilizadas na etapa.
          nullable: false
          items:
            type: string
            nullable: false
          minItems: 3
          maxItems: 10
    CreateStepExternalRequest:
      type: object
      description: Dados para solicitar a crição de um processo seletivo.
      required:
        - title
        - description
      properties:
        title:
          type: string
          description: Título da etapa.
          nullable: false
        description:
          type: string
          description: Descrição da etapa.
          nullable: false
    CreateStepUploadFilesRequest:
      type: object
      description: Dados para solicitar a crição de um processo seletivo.
      required:
        - title
        - description
        - data_upload_files
      properties:
        title:
          type: string
          description: Título da etapa.
          nullable: false
        description:
          type: string
          description: Descrição da etapa.
          nullable: false
        data_upload_files:
          $ref: '#/components/schemas/StepDataUploadFilesRequest'
    StepDataUploadFilesRequest:
      type: array
      description: Informações sobre os arquivos que devem ser enviados na etapa.
      minItems: 1
      maxItems: 3
      nullable: false
      items:
        type: object
        required:
          - description
          - type
        properties:
          description:
            type: string
            description: Descrição de qual arquivo deve ser enviado.
            nullable: false
          type:
            type: string
            description: Tipo do arquivo a ser enviado.
            enum:
              - PDF
              - MP4
            nullable: false
  examples:
...
    CreateStepTheoricalTestRequest:
      summary: "Exemplo de solicitação para criação de uma etapa de teste teórico"
      value:
        title: "Teste teórico de raciocínio lógico"
        description: "Essa é uma etapa para testar o seu raciocínio lógica :)"
        questions_identifiers: [
          "STE-Kamknsd8d",
          "STE-oNKYfc65H",
          "STE-Nuy7H56KJ"
        ]
    CreateStepExternalRequest:
      summary: "Exemplo de solicitação para criação de uma etapa externa"
      value:
        title: "Entrevista com o gestor da vaga"
        description: "Essa é uma etapa para para nos conhecermos melhor e entender as suas expectativas perante a vaga :)"
    CreateStepUploadFilesRequest:
      summary: "Exemplo de solicitação para criação de uma etapa de envio de arquivos"
      value:
        title: "Envio de arquivos de apresentação"
        description: "Essa é uma etapa para você nos informar quem você é :)"
        data_upload_files:
          - description: "currículo"
            type: "PDF"
          - description: "vídeo de apresentação"
            type: "MP4"
...
    InternalErrorResponse:
      summary: "Exemplo de resposta de erro interno no servidor"
      value:
        code: "INTERNAL_ERROR"
        message: "Aconteceu um erro inesperado. Caso o erro persista por um tempo, por favor, entre em contato com o administrador do sistema."
...
    QuestionsInvalidErrorResponse:
      summary: "Exemplo de resposta informando que não é possível cadastrar uma etapa pois foram informados identificadores de questões inválidos"
      value:
        code: "APIX_008"
        message: "Os identificadores fornecidos são inválidos: [QUE-bhjb675Ga, QUE-Ghkb675Ga]."
    QuestionsNotSameTypeErrorResponse:
      summary: "Exemplo de resposta informando que não é possível cadastrar uma etapa pois foram informados identificadores de questões com tipos diferentes. Foram informadas questões discursivas e de múltipla escolha"
      value:
        code: "APIX_009"
        message: "As questões informadas não são do mesmo tipo."
  parameters:
    AuthorizationHeader:
      name: Authorization
      in: header
      description: >-
        Token de acesso no formato JWT.
        
        > Obs.: para mais informações sobre como gerar o token de autenticação, acesso o endpoint '/login'.
      required: true
      schema:
        type: string
...